requirements.txt
torch>=2.0
numpy
tqdm
faiss-cpu
pyyaml
